Backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
dataclasses-json==0.5.14
debugpy==1.6.6
decorator==5.1.1
deep-translator==1.11.4
Django==4.2.3
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==1.2.0
fastjsonschema==2.18.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-trans==2.4.0
google-trans-new==1.1.9
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
inflection==0.5.1
ipykernel==6.21.2
ipython==8.10.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.0.3
jupyter_core==5.2.0
langchain==0.0.266
langsmith==0.0.22
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
nbformat==5.9.2
nest-asyncio==1.5.6
numexpr==2.8.5
numpy==1.25.1
openai==0.27.8
openapi-codec==1.3.2
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
pigar==2.1.1
Pillow==9.5.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.4
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0.1
pyzmq==25.0.0
referencing==0.30.0
requests==2.31.0
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.9.2
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.25.0
streamlit-aggrid==0.3.4.post3
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
yarl==1.9.2
zipp==3.16.2
